+++ src/api/models/__init__.py
@@ -8,7 +8,6 @@ from .base import BaseModel  # noqa: F401
 from .base import Country  # noqa: F401
 from .base import JSONMixin  # noqa: F401
 from .base import Profile  # noqa: F401
-from .base import validate_max_year  # noqa: F401
 from .classification import (  # noqa: F401
     Annotation,
     ClassificationStatus,
