+++ src/api/resources/sampleunitmethods/beltfishmethod.py
@@ -57,7 +57,7 @@ class BeltFishSerializer(BaseAPISerializer):
 
 class ObsBeltFishSerializer(BaseAPISerializer):
     size = serializers.DecimalField(
-        max_digits=5, decimal_places=1, min_value=0, coerce_to_string=False
+        max_digits=5, decimal_places=1, coerce_to_string=False, min_value=0.1
     )
 
     class Meta:
@@ -72,7 +72,7 @@ class ObsBeltFishSerializer(BaseAPISerializer):
         }
 
     def validate_size(self, value):
-        if value is not None and value < 0:
+        if value is not None and value <= 0:
             raise serializers.ValidationError("Size must be greater than or equal to 0.")
         return value
