+++ src/api/resources/sampleunitmethods/benthicpitmethod.py
@@ -54,7 +54,7 @@ class BenthicPITSerializer(BaseAPISerializer):
         max_digits=4,
         decimal_places=2,
         coerce_to_string=False,
-        min_value=0,
+        min_value=0.01,
         max_value=10,
         error_messages={"null": "Interval size is required"},
     )
