+++ src/api/resources/sampleunitmethods/habitatcomplexitymethod.py
@@ -53,7 +53,7 @@ class HabitatComplexitySerializer(BaseAPISerializer):
         max_digits=4,
         decimal_places=2,
         coerce_to_string=False,
-        min_value=0,
+        min_value=0.01,
         max_value=10,
         error_messages={"null": "Interval size is required"},
     )
